instancing

+++ src/renderer/ShaderPipeline.cpp
@@ -22,12 +22,12 @@ void ShaderPipeline::Recompile()
 
 void SetTextures(ShaderPipeline& pipeline, const std::vector<Texture*>& textures)
 {
-    std::vector<uint32_t> samplers;
+    std::vector<int> samplers;
     samplers.reserve(textures.size());
     for (const auto& texture : textures)
     {
         texture->Bind();
-        samplers.push_back(texture->GetImageSlot());
+        samplers.push_back((int)texture->GetImageSlot());
     }
     pipeline.SetUniform("u_texture", samplers.data(), static_cast<uint32_t>(samplers.size()) );
 }
